+++ components/board/genericlinux/board_nonlinux.go
@@ -31,3 +31,8 @@ func RegisterBoard(modelName string, gpioMappings map[int]GPIOBoardMapping, useP
 			},
 		})
 }
+
+// GetGPIOBoardMappings attempts to find a compatible board-pin mapping for the given mappings.
+func GetGPIOBoardMappings(modelName string, boardInfoMappings map[string]BoardInformation) (map[int]GPIOBoardMapping, error) {
+	return nil, errors.New("linux boards are not supported on non-linux OSes")
+}
