+++ components/board/genericlinux/pin_types.go
@@ -1,5 +1,7 @@
 package genericlinux
 
+import "fmt"
+
 // GPIOBoardMapping represents a GPIO pin's location locally within a GPIO chip
 // and globally within sysfs.
 type GPIOBoardMapping struct {
@@ -24,6 +26,27 @@ type PinDefinition struct {
 	PinNameTegraSOC     string
 	PWMChipSysFSDir     string // empty for none
 	PWMID               int    // -1 for none
+}
+
+// board types have to be below so non-linux devices will compile
+
+// BoardInformation details pin definitions and device compatibility for a particular board.
+type BoardInformation struct {
+	PinDefinitions []PinDefinition
+	Compats        []string
+}
+
+// A NoBoardFoundError is returned when no compatible mapping is found for a board during GPIO board mapping.
+type NoBoardFoundError struct {
+	modelName string
+}
+
+func (err NoBoardFoundError) Error() string {
+	return fmt.Sprintf("could not determine %q model", err.modelName)
+}
+
+func noBoardError(modelName string) error {
+	return fmt.Errorf("could not determine %q model", modelName)
 }
 
 // gpioChipData is a struct used solely within GetGPIOBoardMappings and its sub-pieces. It
